requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-SocketIO==5.3.6
google-generativeai==0.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==22.10.2
gevent-websocket==0.10.1
Werkzeug==2.3.7
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Pillow==10.0.1
python-socketio==5.10.0
bidict==0.22.1
groq==0.20.0
Authlib==1.2.1
# Nuevas dependencias para funcionalidades avanzadas
chromadb==0.4.18
sentence-transformers==2.2.2
requests==2.31.0
